Move store_json to streams
